+++ src/main/java/io/andromeda/pippo/routes/fields/MessageField.java
@@ -31,7 +31,7 @@ public class MessageField extends Field {
 
     @Override
     public  boolean validate(Map<String, String> messages) {
-        if (value.equals("")) {
+        if ("".equals(value)) {
             messages.put(alert, Constants.MESSAGE_ID_MESSAGE_EMPTY);
             return false;
         }