+++ src/main/java/io/andromeda/pippo/routes/fields/PhoneField.java
@@ -33,7 +33,7 @@ public class PhoneField extends Field {
 
     @Override
     public  boolean validate(Map<String, String> messages) {
-        if (value.equals("")) {
+        if ("".equals(value)) {
             messages.put(alert, Constants.MESSAGE_ID_PHONE_EMPTY);
             return false;
         } else if (!validatePhoneNumber(value)) {