+++ src/main/java/io/andromeda/pippo/routes/fields/SubjectField.java
@@ -32,7 +32,7 @@ public class SubjectField extends Field {
 
     @Override
     public  boolean validate(Map<String, String> messages) {
-        if (value.equals("")) {
+        if ("".equals(value)) {
             messages.put(alert, Constants.MESSAGE_ID_SUBJECT_EMPTY);
             return false;
         }