+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### [2.0.1] - 2018-05-04
+
+#### Fixed
+- ContactRoute.java was forgotten in the previous release.
+- Improved the checks for empty strings in fields.
+
 ### [2.0.0] - 2018-05-02
 
 #### Fixed